requirements.txt
PyGithub>=1.43.8
tweepy>=3.10.0
